collecting bidirection boundary in single file
