+++ src/shared/shared_ck/particle_dynamics/fluid_dynamics/boundary_condition/bidirectional_boundary_ck.h
@@ -21,21 +21,156 @@
  *                                                                           *
  * ------------------------------------------------------------------------- */
 /**
- * @file 	pressure_boundary_ck.h
+ * @file 	bidirectional_boundary_ck.h
  * @brief 	tbd
  * @author	Xiangyu Hu
  */
 
-#ifndef PRESSURE_BOUNDARY_CK_H
-#define PRESSURE_BOUNDARY_CK_H
+#ifndef BIDIRECTIONAL_BOUNDARY_CK_H
+#define BIDIRECTIONAL_BOUNDARY_CK_H
 
-#include "fluid_boundary_ck.h"
-#include "fluid_boundary_state.h"
+#include "base_body_part.h"
+#include "base_fluid_dynamics.h"
+#include "fluid_boundary_state.hpp"
+#include "particle_operation.hpp"
+#include "particle_reserve.h"
+#include "simple_algorithms_ck.h"
 
 namespace SPH
 {
 namespace fluid_dynamics
 {
+class TagBufferParticlesCK : public BaseLocalDynamics<AlignedBoxPartByCell>
+{
+
+  public:
+    TagBufferParticlesCK(AlignedBoxPartByCell &aligned_box_part);
+    virtual ~TagBufferParticlesCK() {};
+
+    class UpdateKernel
+    {
+      public:
+        template <class ExecutionPolicy, class EncloserType>
+        UpdateKernel(const ExecutionPolicy &ex_policy, EncloserType &encloser);
+        void update(size_t index_i, Real dt = 0.0);
+
+      protected:
+        int part_id_;
+        AlignedBox *aligned_box_;
+        Vecd *pos_;
+        int *buffer_particle_indicator_;
+    };
+
+  protected:
+    int part_id_;
+    SingularVariable<AlignedBox> *sv_aligned_box_;
+    DiscreteVariable<Vecd> *dv_pos_;
+    DiscreteVariable<int> *dv_buffer_particle_indicator_;
+};
+
+template <class TargetPressure>
+class BufferInflowInjectionCK : public BaseLocalDynamics<AlignedBoxPartByCell>
+{
+    using CreateRealParticleKernel = typename SpawnRealParticle::ComputingKernel;
+    using PressureKernel = typename TargetPressure::ComputingKernel;
+
+  public:
+    BufferInflowInjectionCK(AlignedBoxPartByCell &aligned_box_part, ParticleBuffer<Base> &buffer);
+    virtual ~BufferInflowInjectionCK() {};
+
+    class FinishDynamics
+    {
+      public:
+        FinishDynamics(BufferInflowInjectionCK &encloser)
+            : particles_(encloser.particles_), buffer_(encloser.buffer_) {}
+        void operator()() { buffer_.checkEnoughBuffer(*particles_); }
+
+      private:
+        BaseParticles *particles_;
+        ParticleBuffer<Base> &buffer_;
+    };
+
+    class UpdateKernel
+    {
+      public:
+        template <class ExecutionPolicy, class EncloserType>
+        UpdateKernel(const ExecutionPolicy &ex_policy, EncloserType &encloser);
+        void update(size_t index_i, Real dt = 0.0);
+
+      private:
+        int part_id_;
+        AlignedBox *aligned_box_;
+        UnsignedInt *total_real_particles_;
+        CreateRealParticleKernel create_real_particle_;
+        Vecd *pos_;
+        int *buffer_particle_indicator_;
+        PressureKernel pressure_kernel_;
+        Real *physical_time_;
+        Real upper_bound_fringe_;
+    };
+
+  protected:
+    int part_id_;
+    ParticleBuffer<Base> &buffer_;
+    SingularVariable<AlignedBox> *sv_aligned_box_;
+    SingularVariable<UnsignedInt> *sv_total_real_particles_;
+    SpawnRealParticle create_real_particle_method_;
+    DiscreteVariable<Vecd> *dv_pos_;
+    DiscreteVariable<int> *dv_buffer_particle_indicator_;
+    TargetPressure target_pressure_method_;
+    SingularVariable<Real> *sv_physical_time_;
+    Real upper_bound_fringe_;
+};
+
+class BufferOutflowDeletionCK : public BaseLocalDynamics<AlignedBoxPartByCell>
+{
+    using RemoveRealParticleKernel = typename DeleteRealParticle::ComputingKernel;
+
+  public:
+    BufferOutflowDeletionCK(AlignedBoxPartByCell &aligned_box_part);
+    virtual ~BufferOutflowDeletionCK() {};
+
+    class UpdateKernel
+    {
+        struct IsDeletable
+        {
+            int part_id_;
+            AlignedBox *aligned_box_;
+            Vecd *pos_;
+            int *buffer_particle_indicator_;
+            IsDeletable(int part_id, AlignedBox *aligned_box,
+                        Vecd *pos, int *buffer_particle_indicator)
+                : part_id_(part_id), aligned_box_(aligned_box), pos_(pos),
+                  buffer_particle_indicator_(buffer_particle_indicator) {}
+            bool operator()(size_t index_i) const
+            {
+                return buffer_particle_indicator_[index_i] == part_id_ &&
+                       aligned_box_->checkLowerBound(pos_[index_i]);
+            }
+        };
+
+      public:
+        template <class ExecutionPolicy, class EncloserType>
+        UpdateKernel(const ExecutionPolicy &ex_policy, EncloserType &encloser);
+        void update(size_t index_i, Real dt = 0.0); // only works in sequenced policy
+
+      protected:
+        AlignedBox *aligned_box_;
+        Vecd *pos_;
+        IsDeletable is_deltable_;
+        UnsignedInt *total_real_particles_;
+        RemoveRealParticleKernel remove_real_particle_;
+    };
+
+  protected:
+    int part_id_;
+    SingularVariable<AlignedBox> *sv_aligned_box_;
+    SingularVariable<UnsignedInt> *sv_total_real_particles_;
+    DeleteRealParticle remove_real_particle_method_;
+    DiscreteVariable<int> *dv_buffer_particle_indicator_;
+    DiscreteVariable<Vecd> *dv_pos_;
+};
+
 template <class TargetVelocity>
 class VelocityConditionCK : public BaseLocalDynamics<AlignedBoxPartByCell>
 {
@@ -131,4 +266,4 @@ using BidirectionalVelocityBoundaryCK =
     BidirectionalBoundaryCK<ExecutionPolicy, KernelCorrectionType, ZeroGradientPressure, TargetVelocity>;
 } // namespace fluid_dynamics
 } // namespace SPH
-#endif // PRESSURE_BOUNDARY_CK_H
+#endif // BIDIRECTIONAL_BOUNDARY_CK_H
