+++ src/shared/shared_ck/particle_dynamics/fluid_dynamics/boundary_condition/fluid_boundary_ck.h
@@ -115,145 +115,6 @@ class EmitterInflowInjectionCK : public BaseLocalDynamics<AlignedBoxPartType>
     DiscreteVariable<Vecd> *dv_pos_;
     DiscreteVariable<Real> *dv_rho_, *dv_p_;
 };
-
-template <class TargetPressure>
-class BufferInflowInjectionCK : public BaseLocalDynamics<AlignedBoxPartByCell>
-{
-    using CreateRealParticleKernel = typename SpawnRealParticle::ComputingKernel;
-    using PressureKernel = typename TargetPressure::ComputingKernel;
-
-  public:
-    BufferInflowInjectionCK(AlignedBoxPartByCell &aligned_box_part, ParticleBuffer<Base> &buffer);
-    virtual ~BufferInflowInjectionCK() {};
-
-    class FinishDynamics
-    {
-      public:
-        FinishDynamics(BufferInflowInjectionCK &encloser)
-            : particles_(encloser.particles_), buffer_(encloser.buffer_) {}
-        void operator()() { buffer_.checkEnoughBuffer(*particles_); }
-
-      private:
-        BaseParticles *particles_;
-        ParticleBuffer<Base> &buffer_;
-    };
-
-    class UpdateKernel
-    {
-      public:
-        template <class ExecutionPolicy, class EncloserType>
-        UpdateKernel(const ExecutionPolicy &ex_policy, EncloserType &encloser);
-        void update(size_t index_i, Real dt = 0.0);
-
-      private:
-        int part_id_;
-        AlignedBox *aligned_box_;
-        UnsignedInt *total_real_particles_;
-        CreateRealParticleKernel create_real_particle_;
-        Real rho0_;
-        Real sound_speed_;
-        Vecd *pos_;
-        Real *rho_, *p_;
-        int *buffer_particle_indicator_;
-        ConditionKernel condition_;
-        int *previous_surface_indicator_;
-        Real *physical_time_;
-        Real upper_bound_fringe_;
-    };
-
-  protected:
-    int part_id_;
-    ParticleBuffer<Base> &buffer_;
-    SingularVariable<AlignedBox> *sv_aligned_box_;
-    SingularVariable<UnsignedInt> *sv_total_real_particles_;
-    SpawnRealParticle create_real_particle_method_;
-    Real rho0_;
-    Real sound_speed_;
-    DiscreteVariable<Vecd> *dv_pos_;
-    DiscreteVariable<Real> *dv_rho_, *dv_p_;
-    DiscreteVariable<int> *dv_buffer_particle_indicator_;
-    TargetPressure target_pressure_method_;
-    DiscreteVariable<int> *dv_previous_surface_indicator_;
-    SingularVariable<Real> *sv_physical_time_;
-    Real upper_bound_fringe_;
-};
-
-class BufferOutflowDeletionCK : public BaseLocalDynamics<AlignedBoxPartByCell>
-{
-    using RemoveRealParticleKernel = typename DeleteRealParticle::ComputingKernel;
-
-  public:
-    BufferOutflowDeletionCK(AlignedBoxPartByCell &aligned_box_part);
-    virtual ~BufferOutflowDeletionCK() {};
-
-    class UpdateKernel
-    {
-        struct IsDeletable
-        {
-            int part_id_;
-            AlignedBox *aligned_box_;
-            Vecd *pos_;
-            int *buffer_particle_indicator_;
-            IsDeletable(int part_id, AlignedBox *aligned_box,
-                        Vecd *pos, int *buffer_particle_indicator)
-                : part_id_(part_id), aligned_box_(aligned_box), pos_(pos),
-                  buffer_particle_indicator_(buffer_particle_indicator) {}
-            bool operator()(size_t index_i) const
-            {
-                return buffer_particle_indicator_[index_i] == part_id_ &&
-                       aligned_box_->checkLowerBound(pos_[index_i]);
-            }
-        };
-
-      public:
-        template <class ExecutionPolicy, class EncloserType>
-        UpdateKernel(const ExecutionPolicy &ex_policy, EncloserType &encloser);
-        void update(size_t index_i, Real dt = 0.0); // only works in sequenced policy
-
-      protected:
-        AlignedBox *aligned_box_;
-        Vecd *pos_;
-        IsDeletable is_deltable_;
-        UnsignedInt *total_real_particles_;
-        RemoveRealParticleKernel remove_real_particle_;
-    };
-
-  protected:
-    int part_id_;
-    SingularVariable<AlignedBox> *sv_aligned_box_;
-    SingularVariable<UnsignedInt> *sv_total_real_particles_;
-    DeleteRealParticle remove_real_particle_method_;
-    DiscreteVariable<int> *dv_buffer_particle_indicator_;
-    DiscreteVariable<Vecd> *dv_pos_;
-};
-
-class TagBufferParticlesCK : public BaseLocalDynamics<AlignedBoxPartByCell>
-{
-
-  public:
-    TagBufferParticlesCK(AlignedBoxPartByCell &aligned_box_part);
-    virtual ~TagBufferParticlesCK() {};
-
-    class UpdateKernel
-    {
-      public:
-        template <class ExecutionPolicy, class EncloserType>
-        UpdateKernel(const ExecutionPolicy &ex_policy, EncloserType &encloser);
-        void update(size_t index_i, Real dt = 0.0);
-
-      protected:
-        int part_id_;
-        AlignedBox *aligned_box_;
-        Vecd *pos_;
-        int *buffer_particle_indicator_;
-    };
-
-  protected:
-    int part_id_;
-    SingularVariable<AlignedBox> *sv_aligned_box_;
-    DiscreteVariable<Vecd> *dv_pos_;
-    DiscreteVariable<int> *dv_buffer_particle_indicator_;
-};
 } // namespace fluid_dynamics
 } // namespace SPH
 #endif // FLUID_BOUNDARY_CK_H
