+++ src/shared/shared_ck/particle_dynamics/fluid_dynamics/boundary_condition/fluid_boundary_ck.hpp
@@ -80,107 +80,6 @@ void EmitterInflowInjectionCK<AlignedBoxPartType>::UpdateKernel::update(size_t i
     }
 }
 //=================================================================================================//
-template <typename AlignedBoxPartType, class ConditionFunction>
-BufferInflowInjectionCK<AlignedBoxPartType, ConditionFunction>::
-    BufferInflowInjectionCK(AlignedBoxPartType &aligned_box_part, ParticleBuffer<Base> &buffer)
-    : BaseLocalDynamics<AlignedBoxPartType>(aligned_box_part),
-      part_id_(aligned_box_part.getPartID()), buffer_(buffer),
-      sv_aligned_box_(aligned_box_part.svAlignedBox()),
-      sv_total_real_particles_(this->particles_->svTotalRealParticles()),
-      create_real_particle_method_(this->particles_),
-      rho0_(this->particles_->getBaseMaterial().ReferenceDensity()),
-      sound_speed_(0.0),
-      dv_pos_(this->particles_->template getVariableByName<Vecd>("Position")),
-      dv_rho_(this->particles_->template getVariableByName<Real>("Density")),
-      dv_p_(this->particles_->template getVariableByName<Real>("Pressure")),
-      dv_buffer_particle_indicator_(this->particles_->template getVariableByName<int>("BufferParticleIndicator")),
-      condition_function_(this->particles_),
-      dv_previous_surface_indicator_(this->particles_->template getVariableByName<int>("PreviousSurfaceIndicator")),
-      sv_physical_time_(this->sph_system_.template getSystemVariableByName<Real>("PhysicalTime")),
-      upper_bound_fringe_(0.5 * this->sph_body_.getSPHBodyResolutionRef())
-{
-    buffer_.checkParticlesReserved();
-    Fluid &fluid_ = DynamicCast<Fluid>(this, this->particles_->getBaseMaterial());
-    sound_speed_ = fluid_.getSoundSpeed(rho0_);
-}
-//=================================================================================================//
-template <typename AlignedBoxPartType, class ConditionFunction>
-template <class ExecutionPolicy, class EncloserType>
-BufferInflowInjectionCK<AlignedBoxPartType, ConditionFunction>::
-    UpdateKernel::UpdateKernel(const ExecutionPolicy &ex_policy, EncloserType &encloser)
-    : part_id_(encloser.part_id_),
-      aligned_box_(encloser.sv_aligned_box_->DelegatedData(ex_policy)),
-      total_real_particles_(encloser.sv_total_real_particles_->DelegatedData(ex_policy)),
-      create_real_particle_(ex_policy, encloser.create_real_particle_method_),
-      rho0_(encloser.rho0_),
-      sound_speed_(encloser.sound_speed_),
-      pos_(encloser.dv_pos_->DelegatedData(ex_policy)),
-      rho_(encloser.dv_rho_->DelegatedData(ex_policy)),
-      p_(encloser.dv_p_->DelegatedData(ex_policy)),
-      buffer_particle_indicator_(encloser.dv_buffer_particle_indicator_->DelegatedData(ex_policy)),
-      condition_(ex_policy, encloser.condition_function_),
-      previous_surface_indicator_(encloser.dv_previous_surface_indicator_->DelegatedData(ex_policy)),
-      physical_time_(encloser.sv_physical_time_->DelegatedData(ex_policy)),
-      upper_bound_fringe_(encloser.upper_bound_fringe_) {}
-//=================================================================================================//
-template <typename AlignedBoxPartType, class ConditionFunction>
-void BufferInflowInjectionCK<AlignedBoxPartType, ConditionFunction>::
-    UpdateKernel::update(size_t index_i, Real dt)
-{
-    if (!aligned_box_->checkInBounds(pos_[index_i]))
-    {
-        if (aligned_box_->checkUpperBound(pos_[index_i], upper_bound_fringe_) &&
-            buffer_particle_indicator_[index_i] == part_id_ &&
-            index_i < *total_real_particles_)
-        {
-            Vecd original_position = pos_[index_i];
-            UnsignedInt new_particle_index = create_real_particle_(index_i);
-            buffer_particle_indicator_[new_particle_index] = 0;
-            pos_[index_i] = aligned_box_->getUpperPeriodic(original_position);
-            p_[index_i] = condition_(index_i, *physical_time_);
-            rho_[index_i] = p_[index_i] / pow(sound_speed_, 2) + rho0_;
-            previous_surface_indicator_[index_i] = 1;
-        }
-    }
-}
-//=================================================================================================//
-template <class ExecutionPolicy, class EncloserType>
-BufferOutflowDeletionCK::UpdateKernel::
-    UpdateKernel(const ExecutionPolicy &ex_policy, EncloserType &encloser)
-    : aligned_box_(encloser.sv_aligned_box_->DelegatedData(ex_policy)),
-      pos_(encloser.dv_pos_->DelegatedData(ex_policy)),
-      is_deltable_(encloser.part_id_, aligned_box_, pos_,
-                   encloser.dv_buffer_particle_indicator_->DelegatedData(ex_policy)),
-      total_real_particles_(encloser.sv_total_real_particles_->DelegatedData(ex_policy)),
-      remove_real_particle_(ex_policy, encloser.remove_real_particle_method_) {}
-//=================================================================================================//
-void BufferOutflowDeletionCK::UpdateKernel::update(size_t index_i, Real dt)
-{
-    if (!aligned_box_->checkInBounds(pos_[index_i]))
-    {
-        if (is_deltable_(index_i) && index_i < *total_real_particles_)
-        {
-            remove_real_particle_(index_i, is_deltable_);
-        }
-    }
-}
-//=================================================================================================//
-template <class ExecutionPolicy, class EncloserType>
-TagBufferParticlesCK::UpdateKernel::
-    UpdateKernel(const ExecutionPolicy &ex_policy, EncloserType &encloser)
-    : part_id_(encloser.part_id_),
-      aligned_box_(encloser.sv_aligned_box_->DelegatedData(ex_policy)),
-      pos_(encloser.dv_pos_->DelegatedData(ex_policy)),
-      buffer_particle_indicator_(encloser.dv_buffer_particle_indicator_->DelegatedData(ex_policy)) {}
-//=================================================================================================//
-void TagBufferParticlesCK::UpdateKernel::update(size_t index_i, Real dt)
-{
-    if (aligned_box_->checkContain(pos_[index_i]))
-    {
-        buffer_particle_indicator_[index_i] = part_id_;
-    }
-}
-//=================================================================================================//
 } // namespace fluid_dynamics
 } // namespace SPH
 #endif // FLUID_BOUNDARY_CK_HPP
